Update Cobra and pflag vendor libraries
The current version of cobra in our vendor folder has a bug where if
a cobra command has a PersistentPreRunE callback, it expects a
PersistentPostRunE callback, too. If it doesn't exist, then it will
panic.

Adding an empty PersistentPostRunE callback is a workaround, updating
cobra (and pflag as it is a dependency) fixes it.

Change 277 depends on this as it has removed the PersistentPostRunE from
the main command. I think it would be nicer to merge this before merging
that.

Change-Id: I77ebd36a83ccff3076fa9f89a34524dfe42b21eb
Signed-off-by: Julian Carrivick <cjulian@au1.ibm.com>