Remove unnecessary waiting from BDD tests
There are many cases of sleep invocations being made during BDD tests
which are not necessary for the tests to pass. This commit removes
many of the cases of this occurring. I have also reformatted the
tests I have touched to replace tab characters with the standard 4
spaces and make indentation adhere to some semblance of a standard.
Additionally, I have reenabled tests for issue 680 and issue 724 which
pass.

Change-Id: I8695a27f5f35a50b0e84e2c0fddd79dfb885e88a
Signed-off-by: Julian Carrivick <cjulian@au1.ibm.com>