Improve BDDTest decompose times
During BDDTests decomposing takes a lot longer than it should. This is
due to the containers are started using `bash -c "<cmd>"` which does
not pass on signals to the specified command. Hence, docker-compose
sends a SIGTERM which is ignored and after a timeout, sends SIGKILL.
Changing the invocations to `bash -c "exec <cmd>"` ensures the command
receives the signals and terminates in a timely manner.

This change improves BDDTest times by about 5 minutes.

Change-Id: Id0504809740c948758f49ff08961cb377f0a8634
Signed-off-by: Julian Carrivick <cjulian@au1.ibm.com>