Emergency update for busywork
Due to recent changes in APIs etc., busywork does not work with the current
fabric.  In the fabricLogger, changed "uuid" to "txid" to match the new
protos.  In the Tcl library, changed the code to support the new REST API that
now simply takes a single list of base64 arguments for invoke and query.

Change-Id: I78b8f4343294190c404a1e52a926019c8ad52ae1
Signed-off-by: Bishop Brock <bcbrock@us.ibm.com>