Add a container alias map to the bdd context
To simplify the usage of mapping between the docker-compose service
to container name mapping add a dictionary mapping the service (or
alias) to the container data object. Also add a map of full container
name to container data object for convenience.

Standardize the naming of a compose service to containerAlias to
reduce confusion between the actual containerName and the service
defined in docker-compose.
e.g. I deploy chaincode "{chaincodePath}" with ctor "{ctor}" to
     "{containerAlias}"
containerAlias could refer to "vp0" when in reality the container is
called "bddtests_vp0_1"

Rename the ContainerData class to Container and its field:
containerName to name to reduce verbosity in the code base.

Remove functions that were no longer used afer changing the code to
use the map of alias->container.

Signed-off-by: Julian Carrivick <cjulian@au1.ibm.com>
Change-Id: I870719f3ff4db9d07f23d61f1d1a411b6e7ad804