Clarify "not logged in" error
When attempting to deploy with the peer CLI, if the user has not
logged in they get the error "User x not logged in. Use the
'login' command to obtain a security token."

It took me a couple of minutes to figure out what the 'login'
command actually is. I feel the error should specify it as
'peer network login'.

I think this error is specific to the `peer` CLI code,
so it should not affect REST or the Node SDK.

Change-Id: I3d6d4e748d9dbf543c50c2754ef21cacfa14a069
Signed-off-by: Bradley Gorman <bgorman@au1.ibm.com>