Add smart starting of containers during tests
Change 'Given I start peers' to 'Given I start peers, waiting up to
"<time>" seconds for them to be ready', reusing the functions created
for intelligently starting used in 'Given we compose <compose-file>'.

Modify the issue_1874b test to wait for the transaction invoked
during vp0's down time to be committed instead of implicitly
accomplishing this during the wait for 15 seconds statement.

Change-Id: Ib4b00a6d4eb1dfb50f065cba7ac1d4804f447798
Signed-off-by: Julian Carrivick <cjulian@au1.ibm.com>