Revert "Merge "Fix FAB-578" into v0.6"
This reverts commit 562a35fde501d07706cba4b54a1d29068c51f65e, reversing
changes made to 4a78b412ed26cae8757bf34d7f12325a6485e9ef.  This was
done because we inadvertently merged a good chunk of master when
patch a96b9edc80e862 was submitted to the v0.6 branch.  We will
cherry-pick a96b9edc80e862 in a later patch to restore the fix
in the proper way.

Change-Id: Ib16c0220e65697a68a5140960b82780c175d9fa2
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>