FAB-575: New tests (& remote execution capability)
First pass. Adding new tests

This patchset is the code and tests executed for the v0.6
release of fabric. This code includes changes made for the
remote execution of specified behave tests in different
environments.

This includes adding more users to membersrvc for testing
a larger network.

Note changes include updating the directory structure to
accomodate additional test files. The Makefile also needs
to be updated to execute behave tests in the new location.

This structure will change in future versions of fabric.

Change-Id: I3a0fe01d989afadd11366748ecf285335c99c8cc
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>