FAB-860 Fix table crash
GetRows exibits two kinds of errors
  . iterator gets closed asynchronously while chaincode
    is traversing rows
  . channel panic on iterator getting closed (this happens
    when chaincode does not drain the iterator completely)

The two issues are unrelated to each other. Typically the
first issue would kick in and users won't see the second.
For small tables users would see the second if the chaincode
does not read all rows.

First issue is fixed by removing the iterator.Close call in
the GetRows processing. Transaction clean up logic will close
it anyway.

Second issue is fixed by trapping the panic and recovering. This
is based on the fact that the panic can happen only from the
channel close.

NOTE - this is not checked into master yet as current master
       does not completely support Table yet.

Change-Id: I73697b7b5b91d809940c6a969281ecbb8042a763
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>